ModularER_2D/requirements.txt
absl-py==0.9.0
astor==0.8.1
Box2D-kengz==2.3.3
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.2.2
configparser==4.0.2
cycler==0.10.0
deap==1.3.0
future==0.18.2
gast==0.2.2
google-auth==1.8.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
gym==0.15.4
h5py==2.10.0
idna==2.8
kiwisolver==1.1.0
lab==5.5
Markdown==3.1.1
matplotlib==3.1.2
neat-python==0.92
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pip==20.0.1
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pygame==1.9.6
pyglet==1.3.2
PyOpenGL==3.1.5
pyparsing==2.4.5
python-dateutil==2.8.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
setuptools==40.8.0
six==1.13.0
smac==0.12.0
termcolor==1.1.0
tqdm==4.42.1
urllib3==1.25.7
Werkzeug==0.16.0
wheel==0.33.6
wrapt==1.11.2
